Move CharacterSetHeader
